requirements.txt

Flask
flask-cors
Pillow
